requirements.txt
streamlit
yfinance
google-generativeai
python-dotenv
pyyaml
requests
pandas
plotly
pytest
flake8
watchdog
